# Decimal performance improvements