--- a/check.sh
+++ b/check.sh
@@ -19,6 +19,7 @@ cd "$(dirname "$0")"
 (cd scrypto-tests; cargo fmt --check --quiet)
 (cd scrypto-unit; cargo fmt --check --quiet)
 (cd radix-engine; cargo fmt --check --quiet)
+(cd radix-engine-interface; cargo fmt --check --quiet)
 (cd radix-engine-stores; cargo fmt --check --quiet)
 (cd simulator; cargo fmt --check --quiet)
 (cd transaction; cargo fmt --check --quiet)