--- a/radix-engine-interface/src/math/bnum_integer/bits.rs
+++ b/radix-engine-interface/src/math/bnum_integer/bits.rs
@@ -239,8 +239,11 @@ macro_rules! impl_bits {
         )*
     }
 }
-
-impl_bits! { BnumI256, BInt::<4>}
-impl_bits! { BnumI512, BInt::<8>}
-impl_bits! { BnumU256, BUint::<4>}
-impl_bits! { BnumU512, BUint::<8>}
+impl_bits! { BnumI256, BInt::<4> }
+impl_bits! { BnumI384, BInt::<6> }
+impl_bits! { BnumI512, BInt::<8> }
+impl_bits! { BnumI768, BInt::<12> }
+impl_bits! { BnumU256, BUint::<4> }
+impl_bits! { BnumU384, BUint::<6> }
+impl_bits! { BnumU512, BUint::<8> }
+impl_bits! { BnumU768, BUint::<12> }