A set of changes to improve performance for `decimal` and `precise_decimal` types.

New math module has been implemented: `bnum_integer`.
It is strongly inspired by `integer` module, but it is based on [`bnum`](https://crates.io/crates/bnum) which offers:
_Arbitrary precision, fixed-size signed and unsigned integer types for Rust._

General performance improvement is  **95%**
(Detailed results to be provided.)