@lrubasze 2022-12-23
Performance measurements:


Operation | Decimal before | Decimal after | Gain | PreciseDecimal before | PreciseDecimal after | Gain
-- | -- | -- | -- | -- | -- | --
addition | 128.97 ns | 496.31 ps | -99.62% | 192.51 ns | 4.0719 ns | -97.88%
substraction | 99.590 ns | 476.98 ps | -99.52% | 203.51 ns | 3.1508 ns | -98.45%
multiplication | 1.5958 µs | 112.84 ns | -92.93% | 2.3290 µs | 668.78 ns | -71.28%
division | 1.4674 µs | 116.49 ns | -92.06% | 2.3264 µs | 646.91 ns | -72.19%
root | 8.7675 µs | 9.4003 µs | 7.22% | 84.682 µs | 85.380 µs | 0.82%
exponentiation | 7.1623 µs | 507.29 ns | -92.92% | 13.027 µs | 3.2570 µs | -75.00%
conversion from_string | 50.790 µs | 315.38 ns | -99.38% | 150.88 µs | 883.69 ns | -99.41%
conversion to_string | 58.378 µs | 541.70 ns | -99.07% | 240.39 µs | 1.4269 µs | -99.41%