@lrubasze 2023-01-03
> LGTM, although I'm a bit concerned about the maturity of `bnum` (doens't seem to be widely adopted yet).

Thanks for review. Regarding maturity. I agree it is not widely adopted and there are other crates that are more popular.
The thing is they are not fully covering our requirements:
- signed integer types (decimal and precise_decimal is based on it)
- unsigned integer types (they are defined in integer.rs, however not sure if they are really needed..)
- scalable integer size
`bnum` offers that out of the box.

We could have taken more popular/mature crate and reimplement decimal/precise_decimal accordingly, but that would implicate higher regression risk, which I wanted to avoid.